+++ test/test_cmd_select.py
@@ -1,4 +1,6 @@
+import sys
 from   click.testing               import CliRunner
+import pytest
 from   javaproperties_cli.__main__ import javaproperties
 
 INPUT = b'''\
@@ -7,137 +9,127 @@ key = value
 zebra apple
 e\\u00f0=escaped
 e\\\\u00f0=not escaped
-latin-1 = \xF0
+latin-1 = \xC3\xB0
 bmp = \\u2603
 astral = \\uD83D\\uDC10
 bad-surrogate = \\uDC10\\uD83D
 '''
 
-def test_cmd_select_exists():
-    r = CliRunner().invoke(javaproperties, ['select', '-', 'key'], input=INPUT)
-    assert r.exit_code == 0
-    assert r.stdout_bytes == b'#Mon Nov 07 15:29:40 EST 2016\nkey=value\n'
+@pytest.mark.parametrize('args,rc,output', [
+    (
+        ['select', '-', 'key'],
+        0, b'key=value\n',
+    ),
+    (
+        ['select', '-', 'nonexistent'],
+        1, b'javaproperties select: nonexistent: key not found\n',
+    ),
+    (
+        ['select', '-', 'key', 'nonexistent'],
+        1, b'key=value\njavaproperties select: nonexistent: key not found\n',
+    ),
+    (
+        ['select', '--escaped', '-', 'e\\u00F0'],
+        0, b'e\\u00f0=escaped\n',
+    ),
+    (
+        ['select', '--escaped', '--unicode', '-', 'e\\u00F0'],
+        0, b'e\xF0=escaped\n',
+    ),
+    (
+        ['select', '--escaped', '--unicode', '-EUTF-8', '-', 'e\\u00F0'],
+        0, b'e\xC3\xB0=escaped\n',
+    ),
+    (
+        ['select', '--escaped', '-', 'x\\u00F0'],
+        1, b'javaproperties select: x\xC3\xB0: key not found\n'
+    ),
+    (
+        ['select', '-', 'e\\u00f0'],
+        0, b'e\\\\u00f0=not escaped\n',
+    ),
+    (
+        ['select', '-', 'x\\u00f0'],
+        1, b'javaproperties select: x\\u00f0: key not found\n'
+    ),
+    (
+        ['select', '-', b'e\xC3\xB0'],
+        0, b'e\\u00f0=escaped\n',
+    ),
+    (
+        ['select', '--unicode', '-', b'e\xC3\xB0'],
+        0, b'e\xF0=escaped\n',
+    ),
+    (
+        ['select', '--unicode', '-EUTF-8', '-', b'e\xC3\xB0'],
+        0, b'e\xC3\xB0=escaped\n',
+    ),
+    (
+        ['select', '-', b'x\xC3\xB0'],
+        1, b'javaproperties select: x\xC3\xB0: key not found\n',
+    ),
+    (
+        ['select', '-', 'latin-1'],
+        0, b'latin-1=\\u00c3\\u00b0\n',
+    ),
+    (
+        ['select', '--ascii', '-', 'latin-1'],
+        0, b'latin-1=\\u00c3\\u00b0\n',
+    ),
+    (
+        ['select', '-EUTF-8', '-', 'latin-1'],
+        0, b'latin-1=\\u00f0\n',
+    ),
+    (
+        ['select', '--unicode', '-', 'latin-1'],
+        0, b'latin-1=\xC3\xB0\n',
+    ),
+    (
+        ['select', '-EUTF-8', '--unicode', '-', 'latin-1'],
+        0, b'latin-1=\xC3\xB0\n',
+    ),
+    (
+        ['select', '-', 'bmp'],
+        0, b'bmp=\\u2603\n',
+    ),
+    (
+        ['select', '--unicode', '-EUTF-8', '-', 'bmp'],
+        0, b'bmp=\xE2\x98\x83\n',
+    ),
+    (
+        ['select', '-', 'astral'],
+        0, b'astral=\\ud83d\\udc10\n'
+    ),
+    (
+        ['select', '--unicode', '-EUTF-8', '-', 'astral'],
+        0, b'astral=\xF0\x9F\x90\x90\n'
+    ),
+    (
+        ['select', '-', 'bad-surrogate'],
+        0, b'bad-surrogate=\\udc10\\ud83d\n',
+    ),
+    pytest.param(
+        ['select', '--unicode', '-EUTF-8', '-', 'bad-surrogate'],
+        0, b'bad-surrogate=\xED\xB0\x90\xED\xA0\xBD\n',
+        marks=pytest.mark.skipif(
+            sys.version_info[0] != 2,
+            reason='Python 2 only',
+        ),
+    ),
+    pytest.param(
+        ['select', '--unicode', '-EUTF-8', '-', 'bad-surrogate'],
+        0, b'bad-surrogate=\\udc10\\ud83d\n',
+        marks=pytest.mark.skipif(
+            sys.version_info[0] == 2,
+            reason='Python 3 only',
+        ),
+    ),
+])
+def test_cmd_select(args, rc, output):
+    r = CliRunner().invoke(javaproperties, args, input=INPUT)
+    assert r.exit_code == rc, r.stdout_bytes
+    assert r.stdout_bytes == b'#Mon Nov 07 15:29:40 EST 2016\n' + output
 
-def test_cmd_select_not_exists():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', 'nonexistent'
-    ], input=INPUT)
-    assert r.exit_code == 1
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-javaproperties select: nonexistent: key not found
-'''
-
-def test_cmd_select_some_exist():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', 'key', 'nonexistent'
-    ], input=INPUT)
-    assert r.exit_code == 1
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-key=value
-javaproperties select: nonexistent: key not found
-'''
-
-def test_cmd_select_escaped():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '--escaped', '-', 'e\\u00F0'
-    ], input=INPUT)
-    assert r.exit_code == 0
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-e\\u00f0=escaped
-'''
-
-def test_cmd_select_escaped_not_exists():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '--escaped', '-', 'x\\u00F0'
-    ], input=INPUT)
-    assert r.exit_code == 1
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-javaproperties select: x\xC3\xB0: key not found
-'''
-
-def test_cmd_select_not_escaped():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', 'e\\u00f0'
-    ], input=INPUT)
-    assert r.exit_code == 0
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-e\\\\u00f0=not escaped
-'''
-
-def test_cmd_select_not_escaped_not_exists():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', 'x\\u00f0'
-    ], input=INPUT)
-    assert r.exit_code == 1
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-javaproperties select: x\\u00f0: key not found
-'''
-
-def test_cmd_select_utf8():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', b'e\xC3\xB0'  # 'e\u00f0'
-    ], input=INPUT)
-    assert r.exit_code == 0
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-e\\u00f0=escaped
-'''
-
-def test_cmd_select_utf8_not_exists():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', b'x\xC3\xB0'
-    ], input=INPUT)
-    assert r.exit_code == 1
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-javaproperties select: x\xC3\xB0: key not found
-'''
-
-def test_cmd_select_latin1_output():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', 'latin-1'
-    ], input=INPUT)
-    assert r.exit_code == 0
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-latin-1=\\u00f0
-'''
-
-def test_cmd_select_bmp_output():
-    r = CliRunner().invoke(javaproperties, ['select', '-', 'bmp'], input=INPUT)
-    assert r.exit_code == 0
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-bmp=\\u2603
-'''
-
-def test_cmd_select_astral_output():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', 'astral'
-    ], input=INPUT)
-    assert r.exit_code == 0
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-astral=\\ud83d\\udc10
-'''
-
-def test_cmd_select_bad_surrogate_output():
-    r = CliRunner().invoke(javaproperties, [
-        'select', '-', 'bad-surrogate'
-    ], input=INPUT)
-    assert r.exit_code == 0
-    assert r.stdout_bytes == b'''\
-#Mon Nov 07 15:29:40 EST 2016
-bad-surrogate=\\udc10\\ud83d
-'''
-
-# --encoding
 # --outfile
 # --separator
 # -d